pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reinvent4-meta-controller"
version = "0.2.0"
description = "Adaptive exploration/exploitation for de novo molecular design under real-world constraints"
authors = [{name = "REINVENT Meta Team"}]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["drug-discovery", "molecular-design", "reinforcement-learning", "cheminformatics", "qsar"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pandas>=1.5.0",
    "rdkit>=2022.9.1",
    "numpy>=1.23.0",
    "pyyaml>=6.0",
    "scipy>=1.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
r4mc = "r4mc.cli:main"
r4mc-report = "r4mc.report_engine.cli:main"

[tool.setuptools.packages.find]
include = ["r4mc*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.coverage.run]
source = ["r4mc"]
omit = ["*/tests/*", "*/test_*.py"]
